cd/requirements_cd.txt
openai
azureml-core
